refactor(i18n): déduplique le type de langue dans un module langue.ts partagé (12 sites, casts remplacés par parseLangue)
