+++ site/source/hooks/useGetFullURL.ts
@@ -2,8 +2,9 @@ import { useEffect, useRef } from 'react'
 import { useTranslation } from 'react-i18next'
 
 import { useNavigation } from '@/lib/navigation'
+import { AvailableLang, parseLangue } from '@/locales/langue'
 
-const DEVELOPMENT_BASE_PATHS = {
+const DEVELOPMENT_BASE_PATHS: Record<AvailableLang, string> = {
 	fr: '/mon-entreprise',
 	en: '/infrance',
 }
@@ -12,7 +13,7 @@ export const useGetFullURL = () => {
 	const { i18n } = useTranslation()
 	const { currentPath } = useNavigation()
 
-	const language = i18n.language as 'fr' | 'en'
+	const language = parseLangue(i18n.language)
 
 	const pathStart = IS_DEVELOPMENT ? DEVELOPMENT_BASE_PATHS[language] : ''
