+++ site/source/locales/i18n-client.ts
@@ -3,7 +3,8 @@
 import i18next from 'i18next'
 import { initReactI18next } from 'react-i18next'
 
-import { AvailableLang, baseI18nConfig } from './i18nResources'
+import { baseI18nConfig } from './i18nResources'
+import { AvailableLang } from './langue'
 
 export function createI18nClient(langue: AvailableLang) {
 	const instance = i18next.createInstance()
