+++ site/source/locales/i18nResources.ts
@@ -1,3 +1,4 @@
+import { AvailableLang } from './langue'
 import enTranslationsRaw from './ui-en.yaml'
 import unitsTranslationsRaw from './units.yaml'
 
@@ -8,25 +9,6 @@ interface Units {
 
 const enTranslations = enTranslationsRaw as Record<string, string>
 const unitsTranslations = unitsTranslationsRaw as Units
-
-export const SUPPORTED_LANGUAGES = ['fr', 'en'] as const
-export type AvailableLang = (typeof SUPPORTED_LANGUAGES)[number]
-
-export const estLangueSupportée = (valeur: string): valeur is AvailableLang =>
-	(SUPPORTED_LANGUAGES as readonly string[]).includes(valeur)
-
-export const parseLangue = (raw: string | undefined): AvailableLang => {
-	const value = raw ?? 'fr'
-	if (!estLangueSupportée(value)) {
-		throw new Error(
-			`LANGUE invalide : "${value}". Valeurs supportées : ${SUPPORTED_LANGUAGES.join(
-				', '
-			)}.`
-		)
-	}
-
-	return value
-}
 
 export const baseI18nConfig = (langue: AvailableLang) => ({
 	lng: langue,
